+++ src/Event/Dispatcher/DeferringDispatcher.php
@@ -16,10 +16,10 @@ namespace PHPUnit\Event;
  */
 final class DeferringDispatcher implements SubscribableDispatcher
 {
+    use CollectionWindow;
     private readonly SubscribableDispatcher $dispatcher;
     private EventCollection $events;
-    private bool $recording                   = true;
-    private ?EventCollection $collectedEvents = null;
+    private bool $recording = true;
 
     public function __construct(SubscribableDispatcher $dispatcher)
     {
@@ -39,9 +39,7 @@ final class DeferringDispatcher implements SubscribableDispatcher
 
     public function dispatch(Event $event): void
     {
-        if ($this->collectedEvents !== null) {
-            $this->collectedEvents->add($event);
-
+        if ($this->collectDispatchedEvent($event)) {
             return;
         }
 
@@ -52,34 +50,6 @@ final class DeferringDispatcher implements SubscribableDispatcher
         }
 
         $this->dispatcher->dispatch($event);
-    }
-
-    /**
-     * @throws EventsAreAlreadyBeingCollectedException
-     */
-    public function startCollectingEvents(): void
-    {
-        if ($this->collectedEvents !== null) {
-            throw new EventsAreAlreadyBeingCollectedException;
-        }
-
-        $this->collectedEvents = new EventCollection;
-    }
-
-    /**
-     * @throws EventsAreNotBeingCollectedException
-     */
-    public function stopCollectingEvents(): EventCollection
-    {
-        if ($this->collectedEvents === null) {
-            throw new EventsAreNotBeingCollectedException;
-        }
-
-        $events = $this->collectedEvents;
-
-        $this->collectedEvents = null;
-
-        return $events;
     }
 
     public function flush(): void
