+++ src/TextUI/TestRunner.php
@@ -9,13 +9,9 @@
  */
 namespace PHPUnit\TextUI;
 
-use function mt_srand;
-use PHPUnit\Event;
 use PHPUnit\Framework\TestSuite;
 use PHPUnit\Runner\TestRunHistory\TestRunHistory;
-use PHPUnit\Runner\TestSuiteSorter;
 use PHPUnit\TextUI\Configuration\Configuration;
-use Throwable;
 
 /**
  * @no-named-arguments Parameter names are not covered by the backward compatibility promise for PHPUnit
@@ -29,48 +25,14 @@ final class TestRunner
      */
     public function run(Configuration $configuration, TestRunHistory $testRunHistory, TestSuite $suite): void
     {
-        try {
-            Event\Facade::emitter()->testRunnerStarted();
-
-            if ($configuration->executionOrder() === TestSuiteSorter::ORDER_RANDOMIZED) {
-                mt_srand($configuration->randomOrderSeed());
-            }
-
-            $testRunHistory->load();
-
-            if ($configuration->executionOrder() !== TestSuiteSorter::ORDER_DEFAULT ||
-                $configuration->executionOrderDefects() !== TestSuiteSorter::ORDER_DEFAULT ||
-                $configuration->resolveDependencies()) {
-                new TestSuiteSorter($testRunHistory)->reorderTestsInSuite(
-                    $suite,
-                    $configuration->executionOrder(),
-                    $configuration->resolveDependencies(),
-                    $configuration->executionOrderDefects(),
-                );
-
-                Event\Facade::emitter()->testSuiteSorted(
-                    $configuration->executionOrder(),
-                    $configuration->executionOrderDefects(),
-                    $configuration->resolveDependencies(),
-                );
-            }
-
-            (new TestSuiteFilterProcessor)->process($configuration, $suite);
-
-            Event\Facade::emitter()->testRunnerExecutionStarted(
-                Event\TestSuite\TestSuiteBuilder::from($suite),
-            );
-
-            $suite->run();
-
-            Event\Facade::emitter()->testRunnerExecutionFinished();
-            Event\Facade::emitter()->testRunnerFinished();
-        } catch (Throwable $t) {
-            throw new RuntimeException(
-                $t->getMessage(),
-                (int) $t->getCode(),
-                $t,
-            );
-        }
+        new TestRunnerLifecycle()->run(
+            $configuration,
+            $testRunHistory,
+            $suite,
+            static function () use ($suite): void
+            {
+                $suite->run();
+            },
+        );
     }
 }
